requirements.txt
attrs==21.4.0
backcall==0.2.0
cachelib==0.6.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
Flask==2.0.2
Flask-Excel==0.0.7
Flask-Session==0.4.0
idna==3.3
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.3.3
jupyter-client==7.1.0
jupyter-core==4.9.1
lml==0.1.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
nbformat==5.1.3
nest-asyncio==1.5.4
numpy==1.22.0
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
plotly==5.5.0
prompt-toolkit==3.0.24
pyexcel==0.6.7
pyexcel-io==0.6.5
pyexcel-webio==0.1.4
Pygments==2.11.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pywin32==303
pyzmq==22.3.0
requests==2.26.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
spotipy==2.19.0
tenacity==8.0.1
texttable==1.6.4
threadpoolctl==3.0.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
